+++ .clang-format
@@ -1,0 +1,6 @@
+BasedOnStyle: LLVM
+IndentWidth: 4
+UseTab: Never
+MaxEmptyLinesToKeep: 2
+SortIncludes: false
+AlwaysBreakTemplateDeclarations: true
