+++ .gitmodules
@@ -4,3 +4,6 @@
 [submodule "external/tomlplusplus"]
 	path = external/tomlplusplus
 	url = https://github.com/marzer/tomlplusplus.git
+[submodule "external/tclap"]
+	path = external/tclap
+	url = https://git.code.sf.net/p/tclap/code
