+++ CMakeLists.txt
@@ -9,12 +9,15 @@ find_package(webcface CONFIG REQUIRED)
 
 add_subdirectory(external/tiny-process-library)
 add_subdirectory(external/tomlplusplus)
+add_library(tclap INTERFACE)
+target_include_directories(tclap INTERFACE ${CMAKE_CURRENT_LIST_DIR}/external/tclap/include)
 
-add_executable(webcface-launcher launcher/main.cc)
+add_executable(webcface-launcher launcher/main.cc launcher/launcher.cc)
 target_link_libraries(webcface-launcher
     webcface::webcface
     tiny-process-library
     tomlplusplus_tomlplusplus
+    tclap
 )
 install(TARGETS webcface-launcher RUNTIME DESTINATION bin)
