+++ launcher/launcher.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <webcface/webcface.h>
+#include <toml++/toml.hpp>
+
+void launcher(WebCFace::Client &wcli, toml::parse_result &config);
